+++ app/src/main/java/com/example/instabug/gcm/InstabugInstanceIDListenerService.java
@@ -1,0 +1,27 @@
+package com.example.instabug.gcm;
+
+import android.content.Intent;
+
+import com.google.android.gms.iid.InstanceIDListenerService;
+
+/**
+ * Created by vezikon on 6/27/16.
+ */
+public class InstabugInstanceIDListenerService extends InstanceIDListenerService {
+
+    private static final String TAG = "InstabugInstanceIDLS";
+
+    /**
+     * Called if InstanceID token is updated. This may occur if the security of
+     * the previous token had been compromised. This call is initiated by the
+     * InstanceID provider.
+     */
+    // [START refresh_token]
+    @Override
+    public void onTokenRefresh() {
+        // Fetch updated Instance ID token and notify our app's server of any changes (if applicable).
+        Intent intent = new Intent(this, RegistrationIntentService.class);
+        startService(intent);
+    }
+    // [END refresh_token]
+}