+++ app/src/main/java/com/example/instabug/gcm/RegistrationIntentService.java
@@ -1,0 +1,75 @@
+package com.example.instabug.gcm;
+
+import android.app.IntentService;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
+import android.support.v4.content.LocalBroadcastManager;
+import android.util.Log;
+import com.google.android.gms.gcm.GoogleCloudMessaging;
+import com.google.android.gms.iid.InstanceID;
+import com.instabug.library.Instabug;
+
+
+/**
+ * Created by vezikon on 6/27/16.
+ */
+public class RegistrationIntentService extends IntentService {
+
+
+    private static final String TAG = "RegIntentService";
+
+    public RegistrationIntentService() {
+        super(TAG);
+    }
+
+    @Override
+    protected void onHandleIntent(Intent intent) {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+
+        try {
+            // [START register_for_gcm]
+            // Initially this call goes out to the network to retrieve the token, subsequent calls
+            // are local.
+            // R.string.gcm_defaultSenderId (the Sender ID) is typically derived from google-services.json.
+            // See https://developers.google.com/cloud-messaging/android/start for details on this file.
+            // [START get_token]
+            InstanceID instanceID = InstanceID.getInstance(this);
+            String token = instanceID.getToken(QuickstartPreferences.SENDER_ID,
+                    GoogleCloudMessaging.INSTANCE_ID_SCOPE, null);
+            // [END get_token]
+            Log.i(TAG, "GCM Registration Token: " + token);
+
+            // TODO: Implement this method to send any registration to your app's servers.
+            sendRegistrationToServer(token);
+
+            // You should store a boolean that indicates whether the generated token has been
+            // sent to your server. If the boolean is false, send the token to your server,
+            // otherwise your server should have already received the token.
+            sharedPreferences.edit().putBoolean(QuickstartPreferences.SENT_TOKEN_TO_SERVER, true).apply();
+            // [END register_for_gcm]
+        } catch (Exception e) {
+            Log.d(TAG, "Failed to complete token refresh", e);
+            // If an exception happens while fetching the new token or updating our registration data
+            // on a third-party server, this ensures that we'll attempt the update at a later time.
+            sharedPreferences.edit().putBoolean(QuickstartPreferences.SENT_TOKEN_TO_SERVER, false).apply();
+        }
+        // Notify UI that registration has completed, so the progress indicator can be hidden.
+        Intent registrationComplete = new Intent(QuickstartPreferences.REGISTRATION_COMPLETE);
+        LocalBroadcastManager.getInstance(this).sendBroadcast(registrationComplete);
+    }
+
+    /**
+     * Persist registration to third-party servers.
+     *
+     * Modify this method to associate the user's GCM registration token with any server-side account
+     * maintained by your application.
+     *
+     * @param token The new token.
+     */
+    private void sendRegistrationToServer(String token) {
+        // send your registration token to Instabug
+        Instabug.setPushNotificationRegistrationToken(token);
+    }
+
+}